\name{matrix2vector}
\alias{matrix2vector}
\title{convert a matrix to a named vector in which the names are taken from the matrix row and column names}
\usage{
matrix2vector(matrix)
}
\description{
convert a matrix to a named vector in which the names are
taken from the matrix row and column names
}

