\name{rMiscJP-package}
\alias{rMiscJP-package}
\alias{rMiscJP}
\docType{package}
\title{
Miscellaneous R functions
}
\description{
A growing collection of miscellaneous R functions that I find useful.
}
\details{
\tabular{ll}{
Package: \tab rMiscJP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-07\cr
License: \tab GPL 3\cr
}

}
\author{
John R.B. Palmer

Maintainer: John R.B. Palmer <johnrbpalmer@gmail.com>
}
\references{

}
